Split network commands into modules
